Dużo ciekawych rzeczy ;] - konwerter list -> std::set (+dopisek do dokumentacji) - usunięte warningi signed/unsigned poprzez rzutowanie (int) - dostosowanie nazw bibliotek do sconsa - zmienione nazwy TestModule - na Foo, Foobar - drobne poprawki w External

+++ trunk/Source/MapWrapper_py_impl.hpp
@@ -103,7 +103,7 @@ void MapWrapper<Key,V>::print(const Map& x) {
     }
     typename std::map<Key, V>::const_iterator it = x.begin();
 	std::cout << "{";
-	for(int i = 0; i < x.size() - 1; i++)
+	for(int i = 0; i < (int)x.size() - 1; i++)
 	{
         std::cout << it->first << ": " << it->second << ", ";
         ++it;
